- Add performance tests for realistic Sieve iterations
- Add performance tests against glass-pumpkin

`glass-pumpkin` uses more MR tests than required by BPSW, so in the performance tests I used custom method to mimic that behavior. The results (against `crypto-bigint=0.6.0-rc.2`):

1024 bits: 
- `crypto-primes` default: 29ms
- `crypto-primes`, matched with `glass-pumpkin`: 39ms
- `glass-pumpkin`: 46ms

1024 bits, safe prime: 
- `crypto-primes` default: 1.33s
- `crypto-primes`, matched with `glass-pumpkin`: 1.4s
- `glass-pumpkin`: 1.43s