@andrewwhitehead 2024-10-04
Looks like another 10-15% improvement with the `crypto-bigint` master. I see a few places (in the Lucas test) that might benefit from a `Monty::double` method.